@@ -1,9 +1,10 @@
 package com.melisa.innovamotionapp.data.database;
 
+import android.content.Context;
+
 import androidx.room.Database;
 import androidx.room.Room;
 import androidx.room.RoomDatabase;
-import android.content.Context;
 
 @Database(entities = {ReceivedBtDataEntity.class}, version = 1)
 public abstract class InnovaDatabase extends RoomDatabase {
